extract Strategy Selector
